Mood-muffin-final/requirements.txt
flask
flask-socketio
spotipy
requests
python-dotenv
google-generativeai
